refactor(guard): #1218 Step 1b — extract audit sink to app/guard/audit.py
Second module of the three-part decomposition. Same pattern as Step 1a.

Moves from apps/api/app/modules/guard/routers/proxy.py to
apps/api/app/guard/audit.py (verbatim, behavior-identical):
- record()                — was _record_audit (public API, background write)
- _extract_token_counts   — private helper (SSE + JSON usage parsing)
- _estimate_input_tokens  — private helper (block-path token estimate)
- _compute_cost           — private helper (pricing registry lookup)

proxy.py keeps compat re-exports under the original private names. All
six internal call sites in proxy.py unchanged.

Regression harness: 8/8 green against docker Postgres.
Downstream: 239 passed (guard_proxy / guard_pack_matrix / guard_layered_verdict).

Progress:
- Step 1a ✓ app/guard/policy.py    — evaluate() + rule matchers + flatten_prompt()
- Step 1b ✓ app/guard/audit.py     — record() + token/cost helpers
- Step 1c   app/guard/router.py    — upstream fanout (_forward + _stream_chunks)
- Step 2    app/guard/gateway.py   — orchestrator composing policy + audit + router
- Step 3    Lens wired to gateway  — dogfood
